Reorganized file structure and updated README
