+++ app/utils/logger.py
@@ -1,0 +1,10 @@
+import logging
+
+def setup_logger():
+    logging.basicConfig(
+        level=logging.INFO,
+        format="%(asctime)s [%(levelname)s] %(name)s: %(message)s"
+    )
+    return logging.getLogger("sdr_dashboard")
+
+logger = setup_logger()
