--- env
@@ -1,1 +1,0 @@
-GOOGLE_API_KEY = "AIzaSyCk2Y1OLf2qaVfraS9zSTAZ6NR_1QdglPA"
